Spliting files into .cpp
