+++ includes/graphics/spk_window.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#pragma once
-
 #ifdef _WIN32
 	#include "windows/spk_window.hpp"
 #elif __linux__
