+++ includes/network/spk_acceptor.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef _WIN32
+	#include "network/windows/spk_acceptor.hpp"
+#elif __linux__
+	#include "network/linux/spk_acceptor.hpp"
+#endif
